+++ packages/web/viewer/assets/packet-YPE3B663-7rLxk7Hg.js
@@ -1,1 +1,0 @@
-import{p as e}from"./mermaid-parser.core-BYdKYW2k.js";export{e as createPacketServices};