+++ packages/web/viewer/assets/umd-CQV_lJ42.js
@@ -1,2 +1,0 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["./UmdViewer-BNuSd0WH.js","./index-Bs0y_lo-.js","./index-BlC70W4a.css","./UmdViewer-DpObj56a.css"])))=>i.map(i=>d[i]);
-import{C as e,S as t,d as n,t as r}from"./index-Bs0y_lo-.js";var i=e(()=>r(()=>import(`./UmdViewer-BNuSd0WH.js`),__vite__mapDeps([0,1,2,3]),import.meta.url));async function a(e,r){let a=n({render:()=>t(i,{data:e},null)});return a.mount(r),a}export{a as default};